requirements.txt
streamlit
langchain
chromadb
langchain-google-genai
requests
python-dotenv
langchain-community

